feat: move cluster info to overview tab
Closes #623
